Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
  int main () {
     int a,b,c;
-    float avg=(a+b+c)/3;
+    int  avg=(a+b+c)/3;
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
